containers/ai-router/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
httpx>=0.27.0
pyyaml>=6.0.1
pydantic>=2.7.0
